+++ lib/td_dd/grants/policy.ex
@@ -12,10 +12,6 @@ defmodule TdDd.Grants.Policy do
 
   @behaviour Bodyguard.Policy
 
-  def authorize(:cancel_removal, _claims, %{pending_removal: false}), do: false
-
-  def authorize(:request_removal, _claims, %{pending_removal: true}), do: false
-
   def authorize(_action, %{role: "admin"}, _params), do: true
 
   def authorize(:reindex, %{role: "service"}, _params), do: true
@@ -31,18 +27,34 @@ defmodule TdDd.Grants.Policy do
       Permissions.authorized?(claims, :create_grant_request)
   end
 
-  def authorize(:request_removal, %{user_id: user_id} = _claims, %Grant{user_id: user_id}),
-    do: true
-
-  def authorize(:request_removal, %{} = claims, %Grant{data_structure: data_structure}) do
-    Bodyguard.permit?(DataStructures, :request_grant_removal, claims, data_structure)
+  def authorize(
+        :manage_grant_removal_request,
+        %{user_id: user_id} = claims,
+        %Grant{user_id: user_id, data_structure: data_structure}
+      ) do
+    not Bodyguard.permit?(DataStructures, :manage_grant_removal, claims, data_structure)
   end
 
-  def authorize(:cancel_removal, %{user_id: user_id} = _claims, %Grant{user_id: user_id}),
-    do: true
+  def authorize(
+        :manage_grant_removal,
+        %{user_id: user_id} = claims,
+        %Grant{user_id: user_id, data_structure: data_structure}
+      ) do
+    Bodyguard.permit?(DataStructures, :manage_grant_removal, claims, data_structure)
+  end
 
-  def authorize(:cancel_removal, %{} = claims, %Grant{data_structure: data_structure}) do
-    Bodyguard.permit?(DataStructures, :request_grant_removal, claims, data_structure)
+  def authorize(:manage_grant_removal_request, %{} = claims, %Grant{
+        data_structure: data_structure
+      }) do
+    not Bodyguard.permit?(DataStructures, :manage_grant_removal, claims, data_structure) and
+      Bodyguard.permit?(DataStructures, :manage_foreign_grant_removal, claims, data_structure)
+  end
+
+  def authorize(:manage_grant_removal, %{} = claims, %Grant{
+        data_structure: data_structure
+      }) do
+    Bodyguard.permit?(DataStructures, :manage_grant_removal, claims, data_structure) and
+      Bodyguard.permit?(DataStructures, :manage_foreign_grant_removal, claims, data_structure)
   end
 
   def authorize(:update, claims, %Grant{